+++ release_notes.md
@@ -1,3 +1,9 @@
+# added to 'dev' baseline for next release
+**New Features:**
+- Controlled Envelope Single Sideband (CESSB)
+	+ enable via '\cessb on' command
+**Changes:**
+
 # v5.4
 **New Features:**
 - Text console can scroll back to view and work with up to 500 lines of history, in all modes
@@ -14,6 +20,10 @@
 	+ If the `user_style.tpl` file is not found then it will load `default_style.tpl` if it is present.
 	+ If none of these files are located, then it will load the factory style we are all used to.
 	+ There are sample styles in the data folder that can be loaded. Just have to rename one of them to user_style.tpl
+ - Added mode_bal to hw_settings.ini to balance LSB and CWR to USB and CW. 
+	+ 0.8 is a reasonable value. 
+	+ Also harmonized filter band edges.
+	+ removed old ssb_val 
 - Controlled Envelope Single Sideband (CESSB)
 	+ enable via '\cessb on' command